Partial progress on integrating test logger
