+++ scripts/dependencies.props
@@ -4,7 +4,7 @@
     <MSTestVersion>2.0.0</MSTestVersion>
     <NETTestSdkVersion>15.7.2</NETTestSdkVersion>
     <MoqVersion>4.9.0</MoqVersion>
-    <TestLoggerVersion>3.0.26</TestLoggerVersion>
+    <TestLoggerVersion>3.0.31</TestLoggerVersion>
 
     <!-- Test Assets use the minimum supported versions -->
     <NETTestSdkMinimumVersion>15.5.0</NETTestSdkMinimumVersion>
