+++ test/JUnit.Xml.TestLogger.AcceptanceTests/JUnitTestLoggerPathTests.cs
@@ -45,7 +45,7 @@ namespace JUnit.Xml.TestLogger.AcceptanceTests
             string[] expectedResultsFiles = new string[]
             {
                 Path.Combine(DotnetTestFixture.RootDirectory, "JUnit.Xml.TestLogger.NetMulti.Tests.NETFramework46.test-results.xml"),
-                Path.Combine(DotnetTestFixture.RootDirectory, "JUnit.Xml.TestLogger.NetMulti.Tests.NETCoreApp20.test-results.xml")
+                Path.Combine(DotnetTestFixture.RootDirectory, "JUnit.Xml.TestLogger.NetMulti.Tests.NETCoreApp31.test-results.xml")
             };
             foreach (string resultsFile in expectedResultsFiles)
             {
